pyproject.toml
[tool.poetry]
name = "trysms"
version = "0.1.0"
description = ""
authors = ["kofi-teddy <agudogot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
twilio = "^6.62.1"
django-crispy-forms = "^1.12.0"
django-decouple = "^2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
